backend/requirements.txt
Django>=4.2.0

# API and CORS
Django>=4.2.0
 djangorestframework>=3.14.0
 django-cors-headers>=4.3.0

# Data and PDF processing
numpy>=1.21.0
pandas>=1.3.0
PyMuPDF>=1.18.0

# HTTP and APIs
requests>=2.25.0
openai>=1.0.0
pinecone-client>=2.2.0
